moved index.js code to src/Mach.js, documentation
