+++ src/Tree/RootNode.js
@@ -3,9 +3,9 @@
 var Node = require('./Node.js');
 
 /**
-* Represents root of all {@link Tree.Tree}s.
-* @memberof Tree
-*/
+ * Represents root of all {@link Tree.Tree}s.
+ * @memberof Tree
+ */
 class RootNode extends Node {
   constructor() {
     super('ROOT');
