+++ src/Tree/TerminusNode.js
@@ -3,9 +3,9 @@
 var Node = require('./Node.js');
 
 /**
-* Represents the last node in a {@link Tree.Tree};
-* @memberof Tree
-*/
+ * Represents the last node in a {@link Tree.Tree};
+ * @memberof Tree
+ */
 class TerminusNode extends Node {
   constructor() {
     super('TERMINUS');
